File refactoring wip
