+++ src/rust/ide/view/graph-editor/src/component/node.rs
@@ -5,6 +5,7 @@
 
 #[deny(missing_docs)]
 pub mod action_bar;
+pub mod input;
 pub mod port;
 
 pub use port::Expression;
