+++ src/rust/ide/view/graph-editor/src/component/node/input.rs
@@ -1,0 +1,1 @@
+pub mod area;
